requirements.txt
simplejson
requests
pyyaml
pyasesm
